requirements.txt
httpx
python-docx
openai-whisper
